Fix directed memory module placement
