+++ fenra/__init__.py
@@ -1,13 +1,13 @@
 """Fenra package."""
 
-from .directed_memory import (
-    DirectedMemory,
-    DirectedMemoryStore,
-    directed_memories_path,
-    directed_memory_block_for_agent,
-    format_directed_memories_block,
-    get_store,
-)
+import directed_memory as _directed_memory
+
+DirectedMemory = _directed_memory.DirectedMemory
+DirectedMemoryStore = _directed_memory.DirectedMemoryStore
+directed_memories_path = _directed_memory.directed_memories_path
+directed_memory_block_for_agent = _directed_memory.directed_memory_block_for_agent
+format_directed_memories_block = _directed_memory.format_directed_memories_block
+get_store = _directed_memory.get_store
 
 __all__ = [
     "DirectedMemory",
